+++ PyDodo/tests/integration/test_manage_simulation.py
@@ -8,7 +8,6 @@ from pydodo import (
     create_aircraft,
     reset_simulation,
     load_scenario,
-    create_scenario,
     pause_simulation,
     resume_simulation,
     aircraft_position,
@@ -76,7 +75,6 @@ def test_simulation_control():
     resp = set_simulation_rate_multiplier(1.5)
     assert resp == True
 
-
     cmd = set_simulator_mode("agent")
     assert cmd == True
 
@@ -94,11 +92,12 @@ def test_simulation_control():
     pos6 = aircraft_position(aircraft_id)
     assert pos6.loc[aircraft_id]["latitude"] > pos5.loc[aircraft_id]["latitude"]
 
+
 @pytest.mark.skipif(not bb_resp, reason="Can't connect to bluebird")
 @pytest.mark.skipif(not bluesky_sim, reason="Not using BlueSky")
 def test_load_bluesky():
     """
-    Check that can load BlueSky scenario files (if using BlueSky) and change
+    Check that can load BlueSky scenario files (if using BlueSky) and specify
     the sim rate multiplier.
     """
     resp = load_scenario("scenario/8.scn")
@@ -118,43 +117,3 @@ def test_load_fail():
 
     with pytest.raises(AssertionError):
         load_scenario("scenario/8.scn", 0)
-
-
-@pytest.mark.skipif(not bb_resp, reason="Can't connect to bluebird")
-@pytest.mark.skipif(not bluesky_sim, reason="Not using BlueSky")
-def test_create_scenario(rootdir):
-    """
-    Create scenario on the simulator host and load.
-    Check two aircraft created succesfully with associated route.
-    """
-    test_scenario = "dodo-test-scenario"
-
-    test_file = os.path.join(rootdir, test_scenario)
-
-    resp = reset_simulation()
-    assert resp == True
-
-    resp = create_scenario(filename=f"{test_file}.scn", scenario=test_scenario)
-    assert resp == True
-
-    resp = load_scenario(test_scenario)
-    assert resp == True
-
-    time.sleep(1)
-
-    pos = all_positions()
-    assert isinstance(pos, pd.DataFrame)
-    assert len(pos.index) == 2
-
-    aircraft_ids = pos.index
-
-    # access route information for each aircraft
-    route1 = list_route(aircraft_ids[0])
-    route2 = list_route(aircraft_ids[1])
-
-    assert isinstance(route1, pd.DataFrame)
-    assert isinstance(route2, pd.DataFrame)
-    assert len(route1.index) == 3
-    assert len(route2.index) == 3
-    assert len(route1.columns) == 3
-    assert len(route2.columns) == 3
